Added Booking Details screen with RideSharedViewModel
Ran into an issue when trying to pass the Ride object from RideSelectionScreen to the BookingScreen. Finally decided on using a shared ViewModel for the purpose and no ViewModel for the Booking Screen since we are directly using the shared ViewModel. When navigating from the RideSelectionScreen to the BookingScreen, the ride data wasn't persisting correctly because of the way shared state (confirmedRide)was  being managed between screens.